deploy (unicorn) support

+++ lib/kudu/commands/create_project/create_project_base.rb
@@ -32,7 +32,9 @@ module Kudu
       # merge in repo specific settings or use defaults
       @settings.merge!(repository_settings)
 
+      raise InvalidOption, " --patch and --force are not compatible" if options[:force] && options[:patch]
       @force = options[:force]
+      @patch = options[:patch]
       @backup = options[:backup]
       @repo = options[:repo]
     end
@@ -41,7 +43,7 @@ module Kudu
 
     def create_project_skeleton
       target = File.join(Dir.pwd, project_name)
-      if File.exist?(target) && !@force
+      if File.exist?(target) && !@force && !@patch
         Kudu.ui.info "#{target} exists, over-write ? (y/n)"
         overwrite = ''
         while overwrite != 'y' && overwrite != 'n'
@@ -57,11 +59,16 @@ module Kudu
     end
 
     def elaborate(template_file, relative_output_file)
-      template = File.join(Kudu.template_dir, template_file) 
       outfile = File.join(Dir.pwd, project_name, relative_output_file)
+      if @patch && File.exists?(outfile)
+        Kudu.ui.info("File exists: #{outfile} - skipping")
+        return
+      end
+      template = File.join(Kudu.template_dir, template_file) 
       FileUtils.copy(outfile, outfile + ".bu") if @backup && File.exist?(outfile)
       begin
         ErubisInflater.inflate_file_write(template, settings, outfile)
+        Kudu.ui.info("Wrote #{outfile}")
       rescue
         raise TemplateElaborationFailed, "Failed to elaborate template file #{template}"
       end
